古い型やそれを使った関数を削除
