+++ src/lib/autocomplete/generate.ts
@@ -1,28 +1,8 @@
-import { SearchTarget, SearchType } from '../search-target';
-import { NewShip, Ship } from '../ship';
+import { NewShip } from '../ship';
 import { NewAutocompleteOptions } from './option';
 
 const uniqueAndSortTagLabels = (tagLabels: string[]): string[] => {
   return Array.from(new Set(tagLabels)).sort();
-};
-
-export const generateAutocompleteOptions = (
-  ships: Ship[],
-  target: SearchTarget,
-  showAll: boolean
-): string[] => {
-  const shipsShown = ships.filter(({ showDefault }) => showAll || showDefault);
-  switch (target.type) {
-    case SearchType.TAG:
-      return uniqueAndSortTagLabels(
-        shipsShown
-          .flatMap(({ tags }) => tags)
-          .filter(({ category }) => category === target.category)
-          .map(({ label }) => label)
-      );
-    case SearchType.NAME:
-      return shipsShown.map(({ name }) => name);
-  }
 };
 
 export const generateNewAutocompleteOptions = (
