+++ src/lib/filter-ships/search-words.ts
@@ -1,8 +1,3 @@
 import { NewSearchTarget } from '../search-target';
 
-export interface SearchWords {
-  name: string[];
-  tag: string[];
-}
-
 export type NewSearchWords = Record<NewSearchTarget, string[]>;
