+++ src/lib/search-target.ts
@@ -1,0 +1,21 @@
+export const AllSearchTargets = [
+  'names',
+  'categories',
+  'types',
+  'speeds',
+  'ranges',
+  'equipments',
+  'abilities',
+] as const;
+export type NewSearchTarget = typeof AllSearchTargets[number];
+
+export const AllSearchTargetLabels = {
+  names: '名前',
+  categories: '艦種カテゴリ',
+  types: '艦種',
+  speeds: '速力',
+  ranges: '射程',
+  equipments: '装備',
+  abilities: '特性',
+} as const;
+export type SearchTargetLabel = typeof AllSearchTargetLabels[NewSearchTarget];
