--- src/lib/search-target/index.ts
@@ -1,3 +1,0 @@
-export * from './key';
-export * from './label';
-export * from './search-target';
