--- src/lib/search-target/key.ts
@@ -1,5 +1,0 @@
-import { SearchTarget, SearchType } from './search-target';
-
-export const getKeyOfSearchTarget = (target: SearchTarget) => {
-  return target.type === SearchType.NAME ? 'name' : target.category;
-};
