+++ src/lib/ship.ts
@@ -1,16 +1,3 @@
-export interface Tag {
-  category: string;
-  label: string;
-}
-
-export interface Ship {
-  id: number;
-  name: string;
-  kana: string;
-  tags: Tag[];
-  showDefault: boolean;
-}
-
 export interface NewShip {
   id: number;
   name: string;
