--- src/lib/ships-data.ts
@@ -1,5 +1,0 @@
-import { Ship } from './ship';
-
-export interface ShipsData {
-  ships: Ship[];
-}
